WebAppCI&CD/requirements.txt
Flask==2.3.2
azure-storage-blob==12.19.0
